refactor(team): extract office/channels endpoints into broker_office_channels.go
Slice 8. Moves the office-config + channel-management + member-management HTTP
surface (1,576 lines) into a new broker_office_channels.go:
- handleCompany, validateProviderEndpointURL, handleConfig, handleNexRegister
- handleOfficeMembers, handleGenerateMember, handleGenerateChannel
- handleChannels, handleCreateDM, handleChannelMembers
- handleMembers (the formatter-style endpoint)

Relocates 13 office/channel tests from broker_test.go into a new
broker_office_channels_test.go (office roster seeding, channel CRUD,
member lifecycle, channel-slug normalization, default-state load).

Adds two characterization tests:
- TestHandleGenerateMember_RejectsNonPostAndUnauth
- TestHandleCreateDM_RequiresPOST

No behavior change. Total package coverage holds at 63.2%.
